+++ src/1-one/change/api/remove.js
@@ -45,6 +45,7 @@ const methods = {
   /** */
   remove: function (reg) {
     const { indexN } = this.methods.one.pointer
+    this.uncache()
     // two modes:
     //  - a. remove self, from full parent
     let self = this.all()
@@ -76,8 +77,7 @@ const methods = {
       this.ptrs = []
       return self.none()
     }
-    // self._cache = null
-    let res = self.update(ptrs) //return new document
+    let res = self.toView(ptrs) //return new document
     return res
   },
 }