+++ src/3-three/verbs/api/conjugate/toPast.js
@@ -199,7 +199,7 @@ const toPast = function (vb, parsed, form) {
   // console.log(form)
   if (forms.hasOwnProperty(form)) {
     vb = forms[form](vb, parsed)
-    vb.fullSentence().compute(['lexicon', 'preTagger', 'postTagger', 'chunks'])
+    vb.fullSentence().compute(['tagger', 'chunks'])
     return vb
   }
   // do nothing i guess?